# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SeuDrink API",
        "description": "Documentação da API usando Swagger",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
...
        "/api/frutas": {
            "get": {
                "tags": [
                    "Frutas"
                ],
                "summary": "Lista todas as frutas filtradas",
                "operationId": "28b7d822ef44995b335e44536672f012",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Nome da fruta",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de frutas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Fruta"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/fruta/novo": {
            "post": {
                "tags": [
                    "Frutas"
                ],
                "summary": "Cria uma nova fruta",
                "operationId": "3dab7c457257e7aae6479a85e55ba482",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Maçã"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fruta criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Fruta"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/fruta/{id}": {
            "put": {
                "tags": [
                    "Frutas"
                ],
                "summary": "Atualiza uma fruta existente",
                "operationId": "9aa9e732d649dd386da87e8aaa807cc6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da fruta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Pera"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fruta atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Fruta"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Fruta não encontrada"
                    }
                }
            }
        },
        "/api/ingredientes": {
            "get": {
                "tags": [
                    "Ingredientes"
                ],
                "summary": "Lista todos os ingredientes filtrados",
                "operationId": "938da4a3ab271bf6f244f8ac60cf67a4",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Nome do ingrediente",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de ingredientes retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ingrediente"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ingrediente/novo": {
            "post": {
                "tags": [
                    "Ingredientes"
                ],
                "summary": "Cria um novo ingrediente",
                "operationId": "2ac6fc920808513492bc86e6514159da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Açúcar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ingrediente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ingrediente"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ingrediente/{id}": {
            "put": {
                "tags": [
                    "Ingredientes"
                ],
                "summary": "Atualiza um ingrediente existente",
                "operationId": "6f446e09f72b7940fcee922e838465f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do ingrediente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Sal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ingrediente atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ingrediente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ingrediente não encontrado"
                    }
                }
            }
        },
...
        "/api/sugestao/novo": {
            "post": {
                "tags": [
                    "Sugestões"
                ],
                "summary": "Cria uma nova sugestão",
                "operationId": "67b1c17ac3859fc9de67d36d4bc4d21d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "nomeDrink",
                                    "ingredientes",
                                    "recado"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Wilmar"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "wilmar@gmail.com"
                                    },
                                    "nomeDrink": {
                                        "type": "string",
                                        "example": "Mojito"
                                    },
                                    "ingredientes": {
                                        "type": "string",
                                        "example": "Sal, Açucar .."
                                    },
                                    "recado": {
                                        "type": "string",
                                        "example": "Misture tudo e ..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sugestão criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sugestao"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Fruta": {
                "title": "Fruta",
                "description": "Modelo de uma Fruta",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Morango"
                    }
                },
                "type": "object"
            },
            "Ingrediente": {
                "title": "Ingrediente",
                "description": "Modelo de um Ingrediente",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Leite Condensado"
                    }
                },
                "type": "object"
            },
            "Sugestao": {
                "title": "Sugestão",
                "description": "Modelo de uma Sugestao",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Wilmar Filho"
                    },
                    "email": {
                        "type": "email",
                        "example": "wilmarfilho32@hotmail.com"
                    },
                    "nomeDrink": {
                        "type": "string",
                        "example": "Caipirinha"
                    },
                    "ingredientes": {
                        "type": "string",
                        "example": "Açucar, Leite ..."
                    },
                    "recado": {
                        "type": "string",
                        "example": "Esse é um drink muito ..."
                    }
                },
                "type": "object"
            }
        }
    },
...
}